+++ include/clixon_custom.h
@@ -191,3 +191,9 @@
  * To keep the previous behavior (as in 6.0) set this option with #define
  */
 #undef NETCONF_DEFAULT_RETRIEVAL_REPORT_ALL
+
+/*! Development option for RFC 8528 YANG schema mount
+ * Work-in-progress
+ * See also test/test_yang_schema_mount.sh
+ */
+#undef YANG_SCHEMA_MOUNT