+++ lib/src/clixon_validate_minmax.c
@@ -227,7 +227,6 @@ check_unique_list_direct(cxobj     *x,
     char      *str;
     cvec      *cvk;
 
-    cvk = yang_cvec_get(yu);
     /* If list and is sorted by system, then it is assumed elements are in key-order which is optimized
      * Other cases are "unique" constraint or list sorted by user which is quadratic in complexity
      * This second case COULD be optimized if binary insert is made on the vec vector.